% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orecchiette.R
\name{orecchiette}
\alias{orecchiette}
\title{Plot orecchiette}
\usage{
orecchiette(
  data,
  x,
  color_by = x,
  split_by = NULL,
  hide_legend = TRUE,
  title = "Orecchiette",
  x_lab = x,
  y_lab = "Density"
)
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{x}{Name of the outcome variable to plot on x-axis.}

\item{color_by}{Variable name for coloring boxplots (defaults to x)}

\item{split_by}{Optional, variable name for faceting the plot}

\item{hide_legend}{Logical; if TRUE, hides the color legend.}

\item{title}{Character; plot title.}

\item{x_lab}{Character; x-axis label (default is x variable name).}

\item{y_lab}{Character; y-axis label (default is "Density").}
}
\value{
A ggplot or plotly object.
}
\description{
Draws density ("orecchiette") plots for multilevel data.
This works for repeated-measures for example.
Various options are available for color, faceting.
}
\examples{
orecchiette(anthropometry, x='head_circumference', 
            color_by = 'sex', split_by = 'sex~timepoint')
}
