% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigatoni.R
\name{rigatoni}
\alias{rigatoni}
\title{Plot rigatoni}
\usage{
rigatoni(
  data,
  x,
  y,
  color_by = x,
  split_by = NULL,
  split_nrow = NULL,
  split_ncol = NULL,
  hide_legend = TRUE,
  title = "Rigatoni",
  y_lab = y,
  x_lab = x
)
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{x}{Name of the grouping variable.}

\item{y}{Name of the outcome variable to plot on y-axis.}

\item{color_by}{Variable name for coloring boxplots (defaults to x)}

\item{split_by}{Optional, variable name for faceting the plot}

\item{split_nrow}{Optional, number of rows for facet_wrap.}

\item{split_ncol}{Optional, number of columns for facet_wrap.}

\item{hide_legend}{Logical; if TRUE, hides the color legend.}

\item{title}{Character; plot title.}

\item{y_lab}{Character; y-axis label (default is y variable name).}

\item{x_lab}{Character; x-axis label (default is x variable name).}
}
\value{
A ggplot or plotly object.
}
\description{
Draws box ("rigatoni") plots for grouped data.
This works for repeated-measures for example.
Various options are available for color, faceting.
}
\examples{
rigatoni(anthropometry, 
         x = 'nationality', y = 'bmi', split_by='sex')
}
