% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Plot (broken) spaghetti}
\usage{
spaghetti(
  data,
  id,
  y,
  x = "age",
  color_by = NULL,
  split_by = NULL,
  split_ncol = NULL,
  split_nrow = NULL,
  hide_legend = FALSE,
  interactive = FALSE,
  title = "Spaghetti",
  y_lab = y,
  x_lab = x,
  ...
)
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{id}{Name of the grouping variable (e.g. subject ID column in longitudinal data).}

\item{y}{Name of the outcome variable to plot on y-axis.}

\item{x}{Name of the time variable (default is "age").}

\item{color_by}{Optional, variable name for coloring lines/points by group (e.g., sex).}

\item{split_by}{Optional, variable name for faceting the plot (e.g., nationality).}

\item{split_ncol}{Optional, number of columns for facet_wrap.}

\item{split_nrow}{Optional, number of rows for facet_wrap.}

\item{hide_legend}{Logical; if TRUE, hides the color legend.}

\item{interactive}{Logical; if TRUE, returns an interactive plotly plot.}

\item{title}{Character; plot title.}

\item{y_lab}{Character; y-axis label (default is y variable name).}

\item{x_lab}{Character; x-axis label (default is x variable name).}

\item{...}{Additional arguments passed to geom_point().}
}
\value{
A ggplot or plotly object.
}
\description{
Draws subject-wise longitudinal ("spaghetti") plots for grouped data.
This works well for repeated-measures for example, to visualize individual
trajectories.
Various options are available for color, faceting.
The plots can be optionally rendered interactely using \code{plotly}.
Default hover behavior will dispplay the label of grouping (e.g. the
ID of each individual trajectory). This can be useful for instance to identify
outlier subjets.
}
\examples{
# Basic spaghetti plot
spaghetti(anthropometry, y = "head_circumference", id = "id")
# Facet by nationality, color by sex
spaghetti(anthropometry, y = "height", id = "id", color_by = "sex", split_by = "nationality")
# Interactive plot
spaghetti(anthropometry, y = "bmi", id = "id", color_by = "sex", interactive = TRUE)

}
